+++ src/commonMain/kotlin/dev/inmo/krontab/internal/CronDateTime.kt
@@ -1,10 +1,11 @@
 package dev.inmo.krontab.internal
 
-import com.soywiz.klock.DateTime
-import com.soywiz.klock.DateTimeSpan
+import com.soywiz.klock.*
 import dev.inmo.krontab.KronScheduler
 
 /**
+ * @param dayOfweek 0-6
+ * @param year any int
  * @param month 0-11
  * @param dayOfMonth 0-31
  * @param hours 0-23
@@ -12,6 +13,7 @@ import dev.inmo.krontab.KronScheduler
  * @param seconds 0-59
  */
 internal data class CronDateTime(
+    val dayOfweek: Byte? = null,
     val year: Int? = null,
     val month: Byte? = null,
     val dayOfMonth: Byte? = null,
@@ -20,6 +22,7 @@ internal data class CronDateTime(
     val seconds: Byte? = null
 ) {
     init {
+        check(dayOfweek ?.let { it in dayOfWeekRange } ?: true)
         check(year ?.let { it in yearRange } ?: true)
         check(month ?.let { it in monthRange } ?: true)
         check(dayOfMonth ?.let { it in dayOfMonthRange } ?: true)
@@ -29,13 +32,33 @@ internal data class CronDateTime(
     }
 
     internal val klockDayOfMonth = dayOfMonth ?.plus(1)
+    internal val dayOfWeekInt: Int? = dayOfweek ?.toInt()
 }
 
 /**
+ * THIS METHOD WILL <b>NOT</b> TAKE CARE ABOUT [offset] PARAMETER. It was decided due to the fact that we unable to get
+ * real timezone offset from simple [DateTime]
+ *
  * @return The near [DateTime] which happens after [relativelyTo] or will be equal to [relativelyTo]
  */
 internal fun CronDateTime.toNearDateTime(relativelyTo: DateTime = DateTime.now()): DateTime? {
     var current = relativelyTo
+
+    val weekDay = dayOfWeekInt
+    if (weekDay != null && current.dayOfWeek.index0 != weekDay) {
+        do {
+            var diff = weekDay - current.dayOfWeek.index0
+            if (diff < 0) {
+                diff += 7 /* days in week */
+            }
+            current = (current + diff.days).startOfDay
+
+            val next = toNearDateTime(current)
+            if (next == null || next.dayOfWeek.index0 == weekDay) {
+                return next
+            }
+        } while (true)
+    }
 
     seconds?.let {
         val left = it - current.seconds
@@ -77,17 +100,15 @@ internal fun CronDateTime.toNearDateTime(relativelyTo: DateTime = DateTime.now()
     return current
 }
 
-/**
- * @return [KronScheduler] (in fact [CronDateTimeScheduler]) based on incoming data
- */
-internal fun createKronScheduler(
+internal fun createCronDateTimeList(
     seconds: Array<Byte>? = null,
     minutes: Array<Byte>? = null,
     hours: Array<Byte>? = null,
     dayOfMonth: Array<Byte>? = null,
     month: Array<Byte>? = null,
-    years: Array<Int>? = null
-): KronScheduler {
+    years: Array<Int>? = null,
+    weekDays: Array<Byte>? = null
+): List<CronDateTime> {
     val resultCronDateTimes = mutableListOf(CronDateTime())
 
     seconds ?.fillWith(resultCronDateTimes) { previousCronDateTime: CronDateTime, currentTime: Byte ->
@@ -114,5 +135,35 @@ internal fun createKronScheduler(
         previousCronDateTime.copy(year = currentTime)
     }
 
-    return CronDateTimeScheduler(resultCronDateTimes.toList())
+    weekDays ?.fillWith(resultCronDateTimes) { previousCronDateTime: CronDateTime, currentTime: Byte ->
+        previousCronDateTime.copy(dayOfweek = currentTime)
+    }
+
+    return resultCronDateTimes.toList()
 }
+
+/**
+ * @return [KronScheduler] (in fact [CronDateTimeScheduler]) based on incoming data
+ */
+internal fun createKronScheduler(
+    seconds: Array<Byte>? = null,
+    minutes: Array<Byte>? = null,
+    hours: Array<Byte>? = null,
+    dayOfMonth: Array<Byte>? = null,
+    month: Array<Byte>? = null,
+    years: Array<Int>? = null,
+    weekDays: Array<Byte>? = null
+): KronScheduler = CronDateTimeScheduler(createCronDateTimeList(seconds, minutes, hours, dayOfMonth, month, years, weekDays))
+/**
+ * @return [KronScheduler] (in fact [CronDateTimeScheduler]) based on incoming data
+ */
+internal fun createKronSchedulerWithOffset(
+    seconds: Array<Byte>? = null,
+    minutes: Array<Byte>? = null,
+    hours: Array<Byte>? = null,
+    dayOfMonth: Array<Byte>? = null,
+    month: Array<Byte>? = null,
+    years: Array<Int>? = null,
+    weekDays: Array<Byte>? = null,
+    offset: TimezoneOffset
+): KronScheduler = CronDateTimeSchedulerTz(createCronDateTimeList(seconds, minutes, hours, dayOfMonth, month, years, weekDays), offset)