+++ src/commonMain/kotlin/dev/inmo/krontab/internal/CronDateTimeScheduler.kt
@@ -1,7 +1,7 @@
 package dev.inmo.krontab.internal
 
 import com.soywiz.klock.DateTime
-import dev.inmo.krontab.*
+import dev.inmo.krontab.KronScheduler
 import dev.inmo.krontab.collection.plus
 
 /**
@@ -26,8 +26,8 @@ internal data class CronDateTimeScheduler internal constructor(
      *
      * @see toNearDateTime
      */
-    override suspend fun next(relatively: DateTime): DateTime {
-        return cronDateTimes.mapNotNull { it.toNearDateTime(relatively) }.minOrNull() ?: getAnyNext(relatively)
+    override suspend fun next(relatively: DateTime): DateTime? {
+        return cronDateTimes.mapNotNull { it.toNearDateTime(relatively) }.minOrNull()
     }
 }