+++ src/commonTest/kotlin/dev/inmo/krontab/utils/TimeZoneTest.kt
@@ -1,0 +1,27 @@
+package dev.inmo.krontab.utils
+
+import com.soywiz.klock.*
+import dev.inmo.krontab.builder.buildSchedule
+import dev.inmo.krontab.next
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class TimeZoneTest {
+    @Test
+    fun testDifferentTimeZonesReturnsDifferentTimes() {
+        val scheduler = buildSchedule { seconds { every(1) } }
+        val baseDate = DateTime.now().startOfWeek
+        runTest {
+            for (i in 0 until 7) {
+                val now = baseDate + i.days
+                for (j in 0 .. 24) {
+                    val nowTz = now.toOffset(j.hours)
+                    val next = scheduler.next(nowTz)!!
+                    assertEquals(
+                        (nowTz + 1.seconds).utc.unixMillisLong, next.utc.unixMillisLong
+                    )
+                }
+            }
+        }
+    }
+}